Only call git rm when files were deleted

+++ deploy.sh
@@ -97,7 +97,7 @@ if [[ "$BUILD_DIR" = false ]]; then
 		git config --global user.name "10upbot on GitHub"
 
 		# Ensure git archive will pick up any changed files in the directory try.
-    		git rm $(git ls-files --deleted)
+    		test $(git ls-files --deleted) && git rm $(git ls-files --deleted)
   		git add .
   		git commit -m "Include build step changes"
